Supprimer les pages python inutile. Modification du code principale + ajout des tests unitaire
